requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.51.3
pillow>=10.0.0
accelerate>=0.24.0

# Additional utilities
python-pptx>=0.6.21
requests>=2.31.0

# Note: diffusers is installed from git in Dockerfile
# git+https://github.com/huggingface/diffusers.git
